requirements.txt
# requirements.txt
"""
Complete requirements file for Windows environment (no conda)
Install with: pip install -r requirements.txt
"""

# Core Data Science Stack
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.1.0
statsmodels>=0.13.0

# Financial Data & Analysis
yfinance>=0.2.0
alpha-vantage>=2.3.0
quandl>=3.7.0
empyrical-reloaded>=0.5.0
pyfolio-reloaded>=0.9.0
ta>=0.10.0
arch>=5.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
dash>=2.6.0
plotly-dash>=0.1.0

# Technical Analysis
TA-Lib>=0.4.25

# Excel & Reporting
openpyxl>=3.0.0
xlsxwriter>=3.0.0
reportlab>=3.6.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=0.19.0
click>=8.0.0

# Testing & Quality
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# Optional: Professional Data Sources
# bloomberg-api-sdk>=1.0.0  # Uncomment if you have Bloomberg access
# refinitiv-data>=1.0.0     # Uncomment if you have Refinitiv access

# Jupyter for analysis notebooks
jupyter>=1.0.0
jupyterlab>=3.4.0
ipykernel>=6.15.0

# Web server for dashboard
gunicorn>=20.1.0  # For production deployment
